third_party/blink/web_tests/fast/scroll-snap/snaps-after-scrollbar-scrolling-arrow.html

<!DOCTYPE html>
<link rel="help" href="https://drafts.csswg.org/css-scroll-snap-1" />
<link rel="stylesheet" href="resources/simple-snap.css">
<script src="../../resources/testharness.js"></script>
<script src="../../resources/testharnessreport.js"></script>
<script src="../../resources/gesture-util.js"></script>
<script src="../../resources/scrollbar-util.js"></script>
<script src="../../resources/blink-coordinates-util.js"></script>

<div id='scroller'>
  <div id="space"></div>
  <div class="snap left top"></div>
  <div class="snap right top"></div>
  <div class="snap left bottom"></div>
</div>

<script>
  if (window.internals)
    internals.settings.setScrollAnimatorEnabled(false);

  window.onload = async () => {
    // Skip test if there is no visible arrow buttons on the scrollbar.
    if (shouldSkipTests(
            !hasScrollbarArrows(),
            'Scrollbar does not have arrow buttons')) {
      return;
    }

    await waitForCompositorCommit();

    const scroller = document.getElementById("scroller");
    const snapPoint = 400;
    const clickDuration = 100;

    promise_test (async () => {
      await waitForScrollReset(scroller);
      const position = rightArrow(scroller);
      const scrollend = waitForScrollendEvent(scroller);
      await mousePressOn(position.x, position.y, clickDuration);
      await scrollend;
      assert_approx_equals(scroller.scrollLeft, snapPoint, 1);
    }, "Snaps after clicking the forward arrow on the horizontal scrollbar.");

    promise_test (async () => {
      await waitForScrollReset(scroller);
      const position = downArrow(scroller);
      const scrollend = waitForScrollendEvent(scroller);
      await mousePressOn(position.x, position.y, clickDuration);
      await scrollend;
      assert_approx_equals(scroller.scrollTop, snapPoint, 1);
    }, "Snaps after clicking the down arrow on the vertical scrollbar.");
  };
</script>
